feat: Created base structure of the app
Create AppStateVar class to monitor the variables
of the app.  These variables are shared across the
app. These variables may include Language setting,
User Info, etc.

Create AppStateController to manipulate the
variables of AppStateVar. Implementing the
language translation in this controller.

Create App widget that pass the variable down
to the widget tree in flutter. This is done by
using InheritedModel.